requirements.txt
# requirements.txt
flask
firebase_admin
flask_cors
moviepy
spleeter
ffmpeg-python
gunicorn
firebase
google-api-core==1.14.3
google-api-python-client==1.7.11
google-auth==1.7.1
google-auth-httplib2==0.0.3
google-cloud-core==1.0.3
google-cloud-firestore==1.6.0
google-cloud-storage==1.23.0
google-pasta==0.1.8
google-resumable-media==0.5.0
googleapis-common-protos==1.6.0
python_jwt
gcloud
sseclient
asn1crypto==1.0.1
crypto==1.4.1
cryptography==2.7
jwcrypto==0.6.0
pycrypto==2.6.1
requests-toolbelt==0.9.1
ffprobe